--- src/IU/form.java
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="IU">
-<grid id="27dc6" binding="panel1" default-binding="true" layout-manager="GridLayoutManager" row-count="4" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-<margin top="0" left="0" bottom="0" right="0"/>
-<constraints>
-<xy x="20" y="20" width="314" height="175"/>
-</constraints>
-<properties/>
-<border type="none"/>
-<children>
-<component id="de907" class="javax.swing.JButton" binding="bCrear">
-<constraints>
-<grid row="0" column="0" row-span="1" col-span="2" vsize-policy="0" hsize-policy="3" anchor="2" fill="0" indent="0" use-parent-layout="false"/>
-</constraints>
-<properties>
-<background color="-1184256"/>
-<enabled value="true"/>
-<foreground color="-16777216"/>
-<text value="CREAR"/>
-</properties>
-</component>
-<component id="abf02" class="javax.swing.JTextField" binding="tModelo">
-<constraints>
-<grid row="2" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="9" fill="0" indent="0" use-parent-layout="false">
-<preferred-size width="150" height="-1"/>
-</grid>
-</constraints>
-<properties/>
-</component>
-<component id="c260e" class="javax.swing.JTextField" binding="tMatricula">
-<constraints>
-<grid row="2" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="5" fill="0" indent="0" use-parent-layout="false">
-<preferred-size width="150" height="-1"/>
-</grid>
-</constraints>
-<properties/>
-</component>
-<component id="6e3a0" class="javax.swing.JLabel" binding="eModelo">
-<constraints>
-<grid row="1" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="2" fill="0" indent="0" use-parent-layout="false"/>
-</constraints>
-<properties>
-<text value="Modelo"/>
-</properties>
-</component>
-<component id="27f3f" class="javax.swing.JLabel" binding="eMatricula">
-<constraints>
-<grid row="1" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="2" fill="0" indent="0" use-parent-layout="false"/>
-</constraints>
-<properties>
-<text value="Matricula"/>
-</properties>
-</component>
-<component id="bcb34" class="javax.swing.JButton" binding="button1" default-binding="true">
-<constraints>
-<grid row="3" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-</constraints>
-<properties>
-<text value="-"/>
-</properties>
-</component>
-<component id="2b54d" class="javax.swing.JButton" binding="button2" default-binding="true">
-<constraints>
-<grid row="3" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-</constraints>
-<properties>
-<text value="+"/>
-</properties>
-</component>
-</children>
-</grid>
-</form>
